Introduce `RpcParams` in sc-cli
